[fix] 파일업로드 부분 간소화 #208
